+++ demux/demux_raw.c
@@ -237,7 +237,7 @@ static int raw_fill_buffer(demuxer_t *demuxer)
     return 1;
 }
 
-static void raw_seek(demuxer_t *demuxer, float rel_seek_secs, int flags)
+static void raw_seek(demuxer_t *demuxer, double rel_seek_secs, int flags)
 {
     struct priv *p = demuxer->priv;
     stream_t *s = demuxer->stream;
